+++ src/main/java/org/jcodec/api/FrameGrab.java
@@ -84,7 +84,7 @@ public class FrameGrab {
 
         switch (detectFormat) {
         case MOV:
-            MP4Demuxer d1 = new MP4Demuxer(_in);
+            MP4Demuxer d1 = MP4Demuxer.createMP4Demuxer(_in);
             videoTrack = (SeekableDemuxerTrack)d1.getVideoTrack();
             break;
         case MPEG_PS:
